requirements.txt
scikit-learn==1.5.2
pandas==2.2.3
conorm==1.2.0
numpy==2.2.0
matplotlib==3.9.3
python-dotenv==1.0.1
openpyxl==3.1.3
scipy==1.15.1